rename base_loader file as data_bundle
