--- docs/source/fastNLP.io.base_loader.rst
@@ -1,7 +1,0 @@
-fastNLP.io.base\_loader
-=======================
-
-.. automodule:: fastNLP.io.base_loader
-   :members:
-   :undoc-members:
-   :show-inheritance:
