+++ docs/source/fastNLP.io.data_bundle.rst
@@ -1,0 +1,7 @@
+fastNLP.io.data\_bundle
+=======================
+
+.. automodule:: fastNLP.io.data_bundle
+    :members:
+    :undoc-members:
+    :show-inheritance:
