+++ fastNLP/io/embed_loader.py
@@ -9,7 +9,7 @@ import warnings
 import numpy as np
 
 from ..core.vocabulary import Vocabulary
-from .base_loader import BaseLoader
+from .data_bundle import BaseLoader
 from ..core.utils import Option
