+++ fastNLP/io/loader/matching.py
@@ -203,7 +203,8 @@ class QNLILoader(JsonLoader):
         """
         如果您的实验使用到了该数据，请引用
 
-        TODO 补充
+        .. todo::
+            补充
 
         :return:
         """
